requirements.txt
oauth2client==4.1.3
gspread==6.1.2
gspread-dataframe==4.0.0
google-auth==2.35.0
google-api-python-client==2.147.0
pandas==2.2.3
openpyxl==3.1.5
typing_extensions==4.12.2
load-dotenv==0.1.0
python-dotenv==1.0.1
